Move servercomponent tests
